requirements.txt
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
tabula-py>=2.5.0
pdfplumber>=0.9.0
PyPDF2>=3.0.0
